+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Requires macOS 12.0 and higher.
 
 ### Fixed
 - When a user clicked on the `updateDevice` button, the logs would incorrectly state the user was entering the "Unsupported UI" workflow.
+- When running unit tests, Nudge no longer honors the randomDelay key or command line argument
 
 ### Added
 - To ease SOFA testing, you can now pass `-custom-sofa-feed-url` as an argument to override the built-in preferences and/or custom profile/json.
@@ -23,6 +24,9 @@ Requires macOS 12.0 and higher.
 
 ## [2.0.1] - 2024-07-19
 Requires macOS 12.0 and higher.
+
+### Changed
+- The command line argument `-disable-randomDelay` is now `-disable-random-delay`
 
 ### Fixed
 - Some incorrect logic was applied to the `requiredInstallatonDate` when using the new `gracePeriodInstallDelay`